+++ packages/clerk-js/src/ui/components/UserVerification/UVFactorTwoCodeForm.tsx
@@ -1,0 +1,69 @@
+import { useUser } from '@clerk/shared/react';
+import type { __experimental_SessionVerificationResource, PhoneCodeFactor, TOTPFactor } from '@clerk/types';
+import React from 'react';
+
+import type { VerificationCodeCardProps } from '../../elements';
+import { useCardState, VerificationCodeCard } from '../../elements';
+import type { LocalizationKey } from '../../localization';
+import { handleError } from '../../utils';
+import { useAfterVerification } from './use-after-verification';
+
+export type UVFactorTwoCodeCard = Pick<VerificationCodeCardProps, 'onShowAlternativeMethodsClicked'> & {
+  factor: PhoneCodeFactor | TOTPFactor;
+  factorAlreadyPrepared: boolean;
+  onFactorPrepare: () => void;
+  prepare?: () => Promise<__experimental_SessionVerificationResource>;
+};
+
+type SignInFactorTwoCodeFormProps = UVFactorTwoCodeCard & {
+  cardTitle: LocalizationKey;
+  cardSubtitle: LocalizationKey;
+  inputLabel: LocalizationKey;
+  resendButton?: LocalizationKey;
+};
+
+export const UVFactorTwoCodeForm = (props: SignInFactorTwoCodeFormProps) => {
+  const card = useCardState();
+  const { user } = useUser();
+  const { handleVerificationResponse } = useAfterVerification();
+
+  React.useEffect(() => {
+    if (props.factorAlreadyPrepared) {
+      return;
+    }
+
+    void prepare?.();
+  }, []);
+
+  const prepare = props.prepare
+    ? () =>
+        props
+          .prepare?.()
+          .then(() => props.onFactorPrepare())
+          .catch(err => handleError(err, [], card.setError))
+    : undefined;
+
+  const action: VerificationCodeCardProps['onCodeEntryFinishedAction'] = (code, resolve, reject) => {
+    user!
+      .__experimental_verifySessionAttemptSecondFactor({ strategy: props.factor.strategy, code })
+      .then(async res => {
+        await resolve();
+        return handleVerificationResponse(res);
+      })
+      .catch(reject);
+  };
+
+  return (
+    <VerificationCodeCard
+      cardTitle={props.cardTitle}
+      cardSubtitle={props.cardSubtitle}
+      resendButton={props.resendButton}
+      inputLabel={props.inputLabel}
+      onCodeEntryFinishedAction={action}
+      onResendCodeClicked={prepare}
+      safeIdentifier={'safeIdentifier' in props.factor ? props.factor.safeIdentifier : undefined}
+      profileImageUrl={user!.imageUrl}
+      onShowAlternativeMethodsClicked={props.onShowAlternativeMethodsClicked}
+    />
+  );
+};