+++ packages/nextjs/src/client-boundary/uiComponents.tsx
@@ -26,6 +26,7 @@ export {
   SignOutButton,
   SignUpButton,
   UserButton,
+  __experimental_UserVerification,
   GoogleOneTap,
 } from '@clerk/clerk-react';